requirements.txt
# requirements.txt - ДОБАВЛЯЕМ
Flask==2.3.3
langchain==0.0.347
langchain-community==0.0.29
chromadb==0.4.22
sentence-transformers==2.2.2
langchain-huggingface==0.0.3
werkzeug==2.3.7
python-docx==1.1.0
PyPDF2==3.0.1
sqlite3